refactor: admin 페이지 폴더 구조 개선
