+++ src/components/DetailModalHeader.tsx
