+++ src/pages/Admin/Exam/index.tsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from 'react';
 
+import { tableColumnMap, tableColumns } from '@/constants/adminTableColumns';
 import useModal from '@/hooks/useModal';
-import TheoryRegisterModal from '@/pages/Admin/components/TheoryRegisterModal';
+import { ExamTableData } from '@/types/admin';
 
-import { tableColumnMap, tableColumns } from './constants';
-import { TheoryTableData } from './types';
+import ExamRegisterModal from './components/ExamRegisterModal';
 
-function Theory() {
-  const [data, setData] = useState<TheoryTableData[]>([]);
+function Index() {
+  const [data, setData] = useState<ExamTableData[]>([]);
   const { isVisible, toggleModal } = useModal();
 
   useEffect(() => {
@@ -18,13 +18,13 @@ function Theory() {
   return (
     <div className={'w-full h-full flex flex-col gap-20'}>
       <div className={'w-full flex justify-between pr-50'}>
-        <h1 className={'text-30 font-semibold'}>이론 관리</h1>
+        <h1 className={'text-30 font-semibold'}>시험 관리</h1>
         <div className={'flex-row-center gap-30'}>
           <button
             onClick={toggleModal}
             className={'px-10 text-20 border-2 rounded-lg border-neutral-300'}
           >
-            이론 파일 등록
+            시험 파일 등록
           </button>
         </div>
       </div>
@@ -35,9 +35,9 @@ function Theory() {
       >
         <thead className={'bg-slate-300'}>
           <tr className={''}>
-            {tableColumns.theory.map(column => (
+            {tableColumns.exam.map(column => (
               <th key={column} className={'font-medium'}>
-                {tableColumnMap.theory[column]}
+                {tableColumnMap.exam[column]}
               </th>
             ))}
           </tr>
@@ -46,9 +46,9 @@ function Theory() {
           {data.length > 0 ? (
             data.map(row => (
               <tr key={row.id}>
-                {tableColumns.theory.map(column => (
+                {tableColumns.exam.map(column => (
                   <td key={column} className={'text-center'}>
-                    {column === 'file' ? (
+                    {column === 'pdf' || column === 'hwp' ? (
                       <a
                         href={row[column]}
                         target={'_blank'}
@@ -71,9 +71,9 @@ function Theory() {
           )}
         </tbody>
       </table>
-      <TheoryRegisterModal isVisible={isVisible} onClose={toggleModal} />
+      <ExamRegisterModal isVisible={isVisible} onClose={toggleModal} />
     </div>
   );
 }
 
-export default Theory;
+export default Index;
