+++ src/pages/Admin/Theory/components/TheoryRegisterModal.tsx
